init commit introducing pipeline class
