+++ tests/integration/assessments/test_pipeline.py
@@ -1,0 +1,20 @@
+import pytest
+from databricks.labs.remorph.assessments.pipeline import PipelineClass
+
+
+@pytest.fixture(scope="module")
+def pipeline_config():
+    config_path = "path/to/test_config.yaml"
+    return PipelineClass.load_config_from_yaml(config_path)
+
+
+def test_run_pipeline(db_manager, pipeline_config):
+    pipeline = PipelineClass(config=pipeline_config, executor=db_manager)
+    pipeline.execute()
+    assert verify_pipeline_results()
+
+
+def verify_pipeline_results():
+    # Implement the logic to verify the pipeline results
+    # For example, check if the expected data is present in the database
+    return True
